requirements.txt
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
exceptiongroup==1.3.0
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
loguru==0.7.3
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
peewee==3.18.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
sniffio==1.3.1
tomli==2.2.1
typing_extensions==4.13.2
virtualenv==20.31.2
